fixed endTurnIncrement
man this was super annoying lol. the solution was to make it so that changeState calls handleDeath(), but only in combat! 

Also had to modify changeState() to return an object - the clear issue here is that it's changing stuff and then the changes aren't being passed on. In theory, now everything is being passed the correct object. 

(My guess is that this was also why clicking on the buttons returned you to town in previous iterations - the stateObj being passed around was different that state that had been changed by changeState, because the stateObj wasn't being returned by changeState).